Remove if statements from EventForm.jsx
Prior to this commit, there were multiple opening <form> tags
in EventForm.jsx.  For the sake of DRYness, there is now only
one <form> tag.  To accomplish this, I had to remove the if
statements and replace them with react conditional rendering.  The
behavior is the same, but I had to wrap the form-group divs inside
an outer div so that there would only be one top-level element in
the conditionally rendered blocks.