FormModal gets handleFormSubmit function for child
The FormModal component has a primary button that is used to submit
a form.  Prior to this commit, the handleFormSubmit function had
to be passed into the FormModal component.  Now, the FormModal
component reads the handleFormSubmit function from its child
component.  The only child component of the FormModal component
should be a component with the appropriate handleFormSubmit
function bound to the onFormSubmit prop.