Change formTypes to an object, use like an enum
Prior to this commit, formTypes was just a list of strings, each
being one of the valid strings to use in the formType prop of the
Form component.  Now, formTypes is an object that should be treated
like an enumeration of the valid formType values.  I've update the
App component and the tests to work with this change.