If is_private_device, auth_token expires later
I've added the ability for the user to specify if they are using
a public or private device to login or register. If the user selects
that they are on a private device, the created auth token will have
a long expiration date.  If a public device, the auth token will
be set to expire much sooner.  I updated tests and config files
to acommondate this change.