Change decorator check_user_is_admin to users_only
I am in the process of removing the is_admin column from the users
db table.  Protected routes will no longer require users to have
an is_admin value of True.  If someone is a user, they have full
access to all routes.

In the process of making this change, and for this commit, I have
rewritten a decorator so that it only rejects access to a route
if the client is not signed in as a user.  It no longer checks for
an is_admin attribute of the user.