Api gives auth expire time, used for logout timer
The server api now gives the authToken expire time whenever a user
status check, login, or register is successfully completed. The client
uses the expire time to set a timer to automatically log out the user at
that time.  This is used when the user signs in on a public device.