PrivateRoute renders CheckingUserStatus if...
In this commit, I define a new component, CheckingUserStatus, that is
meant to be rendered while the app is checking the status of the user.
The PrivateRoute component renders the CheckingUserStatus component
for as long as the userLoggedIn prop is null.  Now, the PrivateRoute
component will only redirect the user if we know that the client is
not a valid user.  While we don't know the client's user status, we
display the CheckingUserStatus component.

I also added a test to the PrivateRoute.test.jsx file to test that
this behavior works as expected.