Add tests involving Calendar and choiceModal
In this commit, I add two tests to the App.test.jsx file.  The first
test checks that the onCalendarDatesSelect method brings up the
choiceModal if slotInfo.start and slotInfo.end are equal.  The second
test checks that selecting the new job button in the choiceModal brings
up the job form in the formModal, and that the correct date is
automatically filled into the Start Date and End Date input fields.