Render new job event after server creates new job
After submitting a valid new job from the new job form, if the server
successfully adds the new job to the database, the server will return
that new job.  Add the new job to the calendar as a calendar event.
Prior to this commit, the user would have to refresh the page before the
new job would render on the calendar.