Reset job form before showing it to create new job
Prior to this commit, if the user were to click on an existing
job event on the calendar, close the form modal, and then try
to create a new job event, the data from the job that was clicked
before would still be in the job form for creating the new job. Now,
when a user brings up the form modal to create a new job event, the
job form will be reset to its default state.